refactored the performance tracking into a new class + adjusted some logic in Player Supervisor to handle the move
