+++ Assets/ML-Agents/Timers/Training_0_timers.json
@@ -1,1 +1,1 @@
-{"count":1,"self":5.3242584,"total":5.3733528,"children":{"AgentSendState":{"count":27,"self":0.0009947,"total":0.0019979999999999998,"children":{"CollectObservations":{"count":6,"self":0.0010033,"total":0.0010033,"children":null},"CollectDiscreteActionMasks":{"count":6,"self":0,"total":0,"children":null}}},"DecideAction":{"count":27,"self":0.0440942,"total":0.0440942,"children":null},"AgentAct":{"count":27,"self":0.0020019,"total":0.0020019,"children":null}},"gauges":{"Paddle.CumulativeReward":{"count":1,"max":-0.9,"min":-0.9,"runningAverage":-0.9,"value":-0.9,"weightedAverage":-0.9}},"metadata":{"timer_format_version":"0.1.0","start_time_seconds":"1613084040","unity_version":"2019.4.18f1","command_line_arguments":"C:\\Program Files\\Unity\\Hub\\Editor\\2019.4.18f1\\Editor\\Unity.exe -projectpath C:\/Users\/frann\/Documents\/UnityProjects\/CS467\/analysis_logging -useHub -hubIPC -cloudEnvironment production -hubSessionId 97c192c0-6cb8-11eb-8403-7b1263163583 -accessToken JF_-qTPvpTaT6KEZHoEjVv0-dZ1lZ82vmb9zhKD2RT8002f","communication_protocol_version":"1.0.0","com.unity.ml-agents_version":"1.0.6","scene_name":"Training_0","end_time_seconds":"1613084045"}}
+{"count":1,"self":42.553574399999995,"total":43.0526802,"children":{"AgentSendState":{"count":1909,"self":0.045263399999999995,"total":0.0522677,"children":{"CollectObservations":{"count":382,"self":0.0050016999999999995,"total":0.0050016999999999995,"children":null},"CollectDiscreteActionMasks":{"count":382,"self":0.0020025999999999998,"total":0.0020025999999999998,"children":null}}},"DecideAction":{"count":1909,"self":0.4217864,"total":0.4217864,"children":null},"AgentAct":{"count":1909,"self":0.0240481,"total":0.0240481,"children":null}},"gauges":{"Paddle.CumulativeReward":{"count":10,"max":1.00000024,"min":-0.9,"runningAverage":-0.6799999,"value":-0.6,"weightedAverage":-0.674707}},"metadata":{"timer_format_version":"0.1.0","start_time_seconds":"1613087933","unity_version":"2019.4.18f1","command_line_arguments":"C:\\Program Files\\Unity\\Hub\\Editor\\2019.4.18f1\\Editor\\Unity.exe -projectpath C:\/Users\/frann\/Documents\/UnityProjects\/CS467\/analysis_logging -useHub -hubIPC -cloudEnvironment production -hubSessionId 97c192c0-6cb8-11eb-8403-7b1263163583 -accessToken JF_-qTPvpTaT6KEZHoEjVv0-dZ1lZ82vmb9zhKD2RT8002f","communication_protocol_version":"1.0.0","com.unity.ml-agents_version":"1.0.6","scene_name":"Training_0","end_time_seconds":"1613087975"}}
