+++ Assets/Scenes/Training_0.unity
@@ -141,7 +141,7 @@ PrefabInstance:
     - target: {fileID: 3269812750348595291, guid: e863cb15819a52b46a4983fad4a685b2,
         type: 3}
       propertyPath: trackingNumberOfGames
-      value: 1
+      value: 10
       objectReference: {fileID: 0}
     - target: {fileID: 3269812750348595291, guid: e863cb15819a52b46a4983fad4a685b2,
         type: 3}
@@ -159,6 +159,12 @@ PrefabInstance:
       propertyPath: m_BehaviorType
       value: 2
       objectReference: {fileID: 0}
+    - target: {fileID: 8908668329378533745, guid: e863cb15819a52b46a4983fad4a685b2,
+        type: 3}
+      propertyPath: playerData
+      value: 
+      objectReference: {fileID: 11400000, guid: c753f81020ac40b4e96c0a53d1e0d8e8,
+        type: 2}
     - target: {fileID: 9146408817681791003, guid: e863cb15819a52b46a4983fad4a685b2,
         type: 3}
       propertyPath: m_RootOrder
