+++ Assets/Scripts/PerformanceDataManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: fdff6945a9bef3540a63dc59de9a0d02
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
